Moved orientation properties to the SettingsViewModel
