+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/components/drawer/SettingsView.kt
@@ -31,12 +31,10 @@ import org.wycliffeassociates.otter.jvm.workbookapp.ui.components.DeviceCombobox
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.components.LanguageComboboxCell
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.screens.dialogs.AddPluginDialog
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.viewmodel.SettingsViewModel
-import org.wycliffeassociates.otter.jvm.workbookapp.ui.viewmodel.WorkbookDataStore
 import tornadofx.*
 
 class SettingsView : View() {
     private val viewModel: SettingsViewModel by inject()
-    private val workbookDataStore: WorkbookDataStore by inject()
 
     private val addPluginDialog: AddPluginDialog = find()
 
@@ -273,7 +271,7 @@ class SettingsView : View() {
         val successDialog = confirmdialog {
             titleTextProperty.set(messages["settings"])
             messageTextProperty.set(messages["changeLanguageSuccessMessage"])
-            orientationProperty.set(workbookDataStore.orientationProperty.value)
+            orientationProperty.set(viewModel.orientationProperty.value)
 
             cancelButtonTextProperty.set(messages["close"])
             onCloseAction { viewModel.showChangeLanguageSuccessDialogProperty.set(false) }
