+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/screens/RootView.kt
@@ -29,7 +29,7 @@ import org.wycliffeassociates.otter.jvm.workbookapp.ui.OtterApp
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.components.AppBar
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.components.audioerrordialog
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.viewmodel.RootViewModel
-import org.wycliffeassociates.otter.jvm.workbookapp.ui.viewmodel.WorkbookDataStore
+import org.wycliffeassociates.otter.jvm.workbookapp.ui.viewmodel.SettingsViewModel
 import tornadofx.*
 
 class RootView : View() {
@@ -37,7 +37,7 @@ class RootView : View() {
     private val viewModel: RootViewModel by inject()
     private val osThemeDetector = OsThemeDetector.getDetector()
     private val isOSDarkMode = SimpleBooleanProperty(osThemeDetector.isDark)
-    private val workbookDataStore: WorkbookDataStore by inject()
+    private val settingsViewModel: SettingsViewModel by inject()
 
     init {
         // Configure the Workspace: sets up the window menu and external app open events
@@ -66,7 +66,7 @@ class RootView : View() {
         prefWidth = 800.0
         prefHeight = 600.0
 
-        nodeOrientationProperty().bind(workbookDataStore.orientationProperty)
+        nodeOrientationProperty().bind(settingsViewModel.orientationProperty)
 
         borderpane {
             left<AppBar>()
@@ -109,7 +109,7 @@ class RootView : View() {
 
             backgroundImageProperty.set(resources.image("/images/audio_error.png"))
             cancelButtonTextProperty.set(messages["close"])
-            orientationProperty.set(workbookDataStore.orientationProperty.value)
+            orientationProperty.set(settingsViewModel.orientationProperty.value)
 
             errorTypeProperty.bind(viewModel.audioErrorType)
