+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/screens/book/BookSelection.kt
@@ -30,7 +30,7 @@ import org.wycliffeassociates.otter.jvm.controls.dialog.confirmdialog
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.NavigationMediator
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.components.BookCell
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.viewmodel.BookWizardViewModel
-import org.wycliffeassociates.otter.jvm.workbookapp.ui.viewmodel.WorkbookDataStore
+import org.wycliffeassociates.otter.jvm.workbookapp.ui.viewmodel.SettingsViewModel
 import tornadofx.*
 import java.text.MessageFormat
 
@@ -38,7 +38,7 @@ class BookSelection : Fragment() {
 
     private val viewModel: BookWizardViewModel by inject()
     private val navigator: NavigationMediator by inject()
-    private val workbookDataStore: WorkbookDataStore by inject()
+    private val settingsViewModel: SettingsViewModel by inject()
 
     private val breadCrumb = BreadCrumb().apply {
         titleProperty.set(messages["selectBook"])
@@ -68,7 +68,7 @@ class BookSelection : Fragment() {
                     label {
                         addClass("book-wizard__divider")
                         graphic = FontIcon(MaterialDesign.MDI_MENU_RIGHT).apply {
-                            scaleXProperty().bind(workbookDataStore.orientationScaleProperty)
+                            scaleXProperty().bind(settingsViewModel.orientationScaleProperty)
                         }
                     }
                     label {
@@ -131,7 +131,7 @@ class BookSelection : Fragment() {
             backgroundImageFileProperty.bind(viewModel.activeProjectCoverProperty)
             progressTitleProperty.set(messages["pleaseWait"])
             showProgressBarProperty.set(true)
-            orientationProperty.set(workbookDataStore.orientationProperty.value)
+            orientationProperty.set(settingsViewModel.orientationProperty.value)
 
             viewModel.showProgressProperty.onChange {
                 Platform.runLater { if (it) open() else close() }
