+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/screens/dialogs/SplashScreen.kt
@@ -22,14 +22,14 @@ import javafx.geometry.Pos
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.NavigationMediator
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.screens.HomePage
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.styles.SplashScreenStyles
+import org.wycliffeassociates.otter.jvm.workbookapp.ui.viewmodel.SettingsViewModel
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.viewmodel.SplashScreenViewModel
-import org.wycliffeassociates.otter.jvm.workbookapp.ui.viewmodel.WorkbookDataStore
 import tornadofx.*
 
 class SplashScreen : View() {
     private val viewModel: SplashScreenViewModel by inject()
+    private val settingsViewModel: SettingsViewModel by inject()
     private val navigator: NavigationMediator by inject()
-    private val workbookDataStore: WorkbookDataStore by inject()
 
     override val root = stackpane {
         addStylesheet(SplashScreenStyles::class)
@@ -56,7 +56,7 @@ class SplashScreen : View() {
     private fun finish() {
         viewModel.initAudioSystem()
         close()
-        workbookDataStore.setAppOrientation()
+        settingsViewModel.setAppOrientation()
         primaryStage.show()
         navigator.dock<HomePage>()
     }
